+++ parcel.go
@@ -8,6 +8,24 @@ type ParcelStore struct {
 	db *sql.DB
 }
 
+func NewParcelStore(db *sql.DB) ParcelStore {
+	return ParcelStore{db: db}
+}
+
+const (
+	ParcelStatusRegistered = "registered"
+	ParcelStatusSent       = "sent"
+	ParcelStatusDelivered  = "delivered"
+)
+
+type Parcel struct {
+	Number    int
+	Client    int
+	Status    string
+	Address   string
+	CreatedAt string
+}
+
 func (s ParcelStore) Add(p Parcel) (int, error) {
 	query := `INSERT INTO parcel (client, status, address, created_at) VALUES (?, ?, ?, ?)`
 	result, err := s.db.Exec(query, p.Client, p.Status, p.Address, p.CreatedAt)
